Route the standard go logger and stdout/stderr to the logrus logger
Some third-party modules and shared libraries will occasionally produce
output on stdout/stderr. Such output would go unformatted into the log
file.

This commit ensures that all such output is sent to the logger and
formatted so that it gets timestamps and an initial "stdlog", "stdout",
or "stderr" column that denotes the origin.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>