Make Test_RestartInterceptedPod work with the FTP client
The fuse client isn't terminated when the intercepted pod changes.
Instead, we just set a new remote address for the FTP client. This means
that the fuse client's cache is still valid for a short period of time.

This commit ensures that the test will access an entry that wasn't
cached.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>